Move `ValidationError` to new `validation` module
