+++ crates/fj-operations/src/group.rs
@@ -1,7 +1,6 @@
 use fj_interop::debug::DebugInfo;
 use fj_kernel::{
-    algorithms::Tolerance,
-    shape::{Shape, ValidationError},
+    algorithms::Tolerance, shape::Shape, validation::ValidationError,
 };
 use fj_math::Aabb;
