+++ crates/fj-operations/src/lib.rs
@@ -27,8 +27,7 @@ mod transform;
 
 use fj_interop::debug::DebugInfo;
 use fj_kernel::{
-    algorithms::Tolerance,
-    shape::{Shape, ValidationError},
+    algorithms::Tolerance, shape::Shape, validation::ValidationError,
 };
 use fj_math::Aabb;
